requirements.txt
python-dotenv
pandas
numpy
certifi
pymongo[srv]==4.7
scikit-learn
mlflow
pyaml
dagshub
fastapi
uvicorn
python-multipart
joblib
python-whois
streamlit

# -e .
